+++ src/dmd_era5/__init__.py
@@ -9,7 +9,7 @@ from importlib.metadata import version
 from dmd_era5.config_reader import config_reader
 from dmd_era5.logger import log_and_print, setup_logger
 from dmd_era5.create_mock_data import create_mock_era5
-from dmd_era5.era5_processing import slice_era5_dataset, thin_era5_dataset, standardize_data
+from dmd_era5.slice_tools import slice_era5_dataset, thin_era5_dataset, standardize_data
 
 __all__ = [
     "config_reader",
