+++ src/dmd_era5/slice_tools/__init__.py
@@ -1,4 +1,4 @@
-from dmd_era5.era5_processing.era5_processing import (
+from dmd_era5.slice_tools.slice_tools import (
     slice_era5_dataset,
     thin_era5_dataset,
     standardize_data,
